Makefiles

+++ Libraries/UPnP/UPnP.h
@@ -66,7 +66,7 @@ private:
 	std::string _udn;
 	std::string _st;
 	std::map<int32_t, Packets> _packets;
-	int64_t _lastAdvertisement;
+	int32_t _lastAdvertisement;
 
 	void getAddress();
 	void getUDN();
